reply实现

+++ src/main/java/com/bbs/dao/adminDao.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
  * Created by valar on 2019/1/5.
  */
 @Repository
-public interface adminDao extends CrudRepository<admin,Long> {
+public interface adminDao extends CrudRepository<admin,Integer> {
     @Query("select s from admin s where s.name=?1 and s.pw=?2")
     admin findByAdminNameAndPw(String name,String pw);
 }
